[Gradient Compression] Replace the assertions in PowerSGD comm hook by stream syncrhonization
Previously the assertion that prevents illegal memory access is because of the torch.any that returns a boolean value, which initiates a data transfer from the device to the host and forces a synchronization.

An explicit synchronization is more to the point.

Original PR issue: Investigate Applying PowerSGD to Communication Hook for Gradient Compression #47202

Differential Revision: [D25573484](https://our.internmc.facebook.com/intern/diff/D25573484/)

ghstack-source-id: 118664204
Pull Request resolved: https://github.com/pytorch/pytorch/pull/49435